Rename GeoTiffReader to RasterReader
